test 2350

+++ src/main/java/dto/TransferRequestDto.java
@@ -1,12 +1,14 @@
 package dto;
 
+import java.math.BigDecimal;
+
 public class TransferRequestDto {
 
-    public final double amount;
+    public final BigDecimal amount;
     public final String fromId;
     public final String toId;
 
-    public TransferRequestDto(double amount, String fromId, String toId) {
+    public TransferRequestDto(BigDecimal amount, String fromId, String toId) {
         this.amount = amount;
         this.fromId = fromId;
         this.toId = toId;
